requirements.txt
bilibili-api-python
requests
pandas
jieba
matplotlib
pyecharts
wordcloud
tqdm
configparser
asyncio
streamlit
httpx 